Refactor printing to allow for custom changes to lua filter
